styleTTS/requirements.txt
numpy>=1.24
soundfile>=0.12
huggingface_hub>=0.24
pyyaml>=6.0
